fix mermaid
